+++ app/Http/Controllers/ArticleController.php
@@ -1,10 +1,11 @@
 <?php
-
 namespace App\Http\Controllers;
 
 use App\Models\Article;
 use App\Models\CategoryArticle;
 use Illuminate\Http\Request;
+use Backpack\NewsCRUD\app\Models\Tag;
+use DB;
 
 class ArticleController extends Controller
 {
@@ -28,23 +29,35 @@ class ArticleController extends Controller
                                         ->limit(8)
                                         ->get();
 
+        $recommendedTutorials = Article::where('category_id', '!=', $news->id)
+                                        ->limit(8)
+                                        ->get();                                
+
        $tutorialCategory = CategoryArticle::whereSlug('tutorials')->get()->first();
 
        return view('articles.news.content', [
            'news' => $news,
            'recommendedNews' => $recommendedNews,
-           'tutorialCategory' => $tutorialCategory
+           'tutorialCategory' => $tutorialCategory,
+           'recommendedTutorials' => $recommendedTutorials,
        ]);
     }
 
 
-    public function tutorials($slug)
+    public function tutorials()
     {
-        $category = CategoryArticle::whereSlug($slug)->get()->first();
-        $tutorials = $category->children;
-
-        return view('articles.tutorials.tutorials_in_category', [
+        $tutorialCategory = CategoryArticle::where('slug', 'tutorials')->get()->first();
+      
+        $tutorials = DB::table('articles')
+                        ->select(['*', 'categories.slug as cat_slug', 'articles.slug as slug'])
+                        ->join('categories', function($join) use ($tutorialCategory) {
+                                $join->on('categories.id', '=', 'articles.category_id')
+                                ->where('categories.parent_id', $tutorialCategory->id);
+            })->paginate(15);
+        
+        return view('articles.tutorials.index', [
             'tutorials' => $tutorials,
+            'tutorialCategory' => $tutorialCategory,
         ]);
     }
 
@@ -57,11 +70,19 @@ class ArticleController extends Controller
     {
         $category = CategoryArticle::whereSlug($slug)->get()->first();
 
-        $tutorials = $category->articles->chunk(2);
+        $tutorials = DB::table('articles')
+                        ->select(['*', 'categories.slug as cat_slug', 'articles.slug as slug'])
+                        ->join('categories', function($join) use ($category) {
+                            $join->on('categories.id', '=', 'articles.category_id')
+                            ->where('categories.id', $category->id);
+            })->paginate(15);
 
-        return view('articles.tutorials.tutorials_in_category', [
+        $tutorialCategory = CategoryArticle::whereSlug('tutorials')->get()->first();
+
+        return view('articles.tutorials.category', [
             'category' => $category,
             'tutorials' => $tutorials,
+            'tutorialCategory' => $tutorialCategory,
         ]);
 
     }
@@ -71,8 +92,56 @@ class ArticleController extends Controller
      */
     public function tutorial(Request $request)
     {
+
+        //Tutorial wich getted by slug
         $tutorial = Article::whereSlug($request->slug)->get()->first();
 
+
+        //The tutorials in the same category
+        $reletedTutotrials = Article::where('id', '!=', $tutorial->id)
+                            ->where('category_id', $tutorial->category_id)
+                            ->orderBy('created_at', 'desc')
+                            ->limit(8)
+                            ->get();
+
+        //Another tutorials for more                    
+        $recommendedTutorials = Article::where('id', '!=', $tutorial->id)
+                                    ->where('category_id', '!=', $tutorial->category_id)
+                                    ->orderBy('created_at', 'desc')
+                                    ->limit(8)
+                                    ->get();
+
+        //Category of tutorials                            
+        $tutorialCategory = CategoryArticle::whereSlug('tutorials')->get()->first();
+
+        return view('articles.tutorials.tutorial', [
+            'tutorial' => $tutorial,
+            'reletedTutotrials' => $reletedTutotrials,
+            'recommendedTutorials' => $recommendedTutorials,
+            'tutorialCategory' => $tutorialCategory
+        ]);
     }
+
+
+    public function tags(Request $request) 
+    {       
+        $tag = Tag::whereSlug($request->slug)->get()->first();
+        
+
+        //Category of tutorials                            
+        $tutorialCategory = CategoryArticle::whereSlug('tutorials')->get()->first();
+
+        //ids of tutorials
+        $ids = $tutorialCategory->children->pluck('id');
+
+        $tutorials = $tag->articles()->whereIn('category_id', $ids)->get();
+
+        return view('articles.tag', [
+            'tag' => $tag,
+            'tutorials' => $tutorials,
+            'tutorialCategory' => $tutorialCategory,
+        ]);
+    }
+
 
 }